+++ public/js/controllers/returnController.js
@@ -7,8 +7,8 @@
  */
 
 angular.module('BibBox').controller('ReturnController', [
-  '$scope', '$controller', '$location', '$timeout', 'userService', 'receiptService', 'config', '$modal',
-  function ($scope, $controller, $location, $timeout, userService, receiptService, config, $modal) {
+  '$scope', '$controller', '$location', '$timeout', 'userService', 'receiptService', 'config', '$modal', 'loggerService',
+  function ($scope, $controller, $location, $timeout, userService, receiptService, config, $modal, loggerService) {
     'use strict';
 
     // Instantiate/extend base controller.
@@ -148,9 +148,9 @@ angular.module('BibBox').controller('ReturnController', [
           function error(err) {
             $scope.baseResetIdleWatch();
 
-            console.error('Return error', err);
+            loggerService.error('Check-in: ' + err);
 
-            for (i = 0; i < $scope.materials.length; i++) {
+            for (var i = 0; i < $scope.materials.length; i++) {
               if ($scope.materials[i].id === material.id) {
                 material = $scope.materials[i];
 
@@ -287,8 +287,7 @@ angular.module('BibBox').controller('ReturnController', [
           // Ignore.
         },
         function (err) {
-          // @TODO: Report error to user.
-          console.log(err);
+          loggerService.error(err);
         }
       );