+++ public/js/services/userService.js
@@ -8,13 +8,16 @@
  */
 
 
-angular.module('BibBox').service('userService', ['$q', '$timeout', '$location', 'config', 'proxyService', 'userTrackerService',
-  function ($q, $timeout, $location, config, proxyService, userTrackerService) {
+angular.module('BibBox').service('userService', ['$q', '$timeout', '$location', 'config', 'proxyService', 'userTrackerService', 'loggerService',
+  function ($q, $timeout, $location, config, proxyService, userTrackerService, loggerService) {
     'use strict';
 
     this.username = null;
     this.password = null;
     this.loggedIn = false;
+
+
+    proxyService.emit('logger.frontend', 'TEST TEST');
 
     /**
      * Is user logged in?
@@ -92,8 +95,7 @@ angular.module('BibBox').service('userService', ['$q', '$timeout', '$location', 
               deferred.reject(err);
             },
             function () {
-              // @TODO: What to do...
-              console.log(err);
+              loggerService.error(err);
             });
           }
           else {