+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed how modals are hidden.
 * Added user logout for reservations and status.
 * Changed receipt buttons for borrow/return, so they are only available when material has been accepted by FBS.
+* Added logger service to send front-end message to the backend.
 
 ### v1.3.0